chore(plugins): sync plugin updates from OVH (19 plugins)
Pulled latest plugin versions installed on prod (mon-pilates.bzh) so the
git tree no longer lags behind. Avoids the CI rsync from overwriting newer
plugin code with older cached versions.

Major version bumps :
- akismet 5.6 → 5.7
- backwpup 5.6.8 → 5.6.9
- complianz-gdpr 7.4.5 → 7.4.6
- facebook-for-woocommerce 3.6.2 → 3.6.3
- jetpack 15.7 → 15.8
- klaviyo 3.7.3 → 3.7.4
- limit-login-attempts-reloaded 3.1.0 → 3.2.1 (major)
- megamenu 3.8.1 → 3.9.2.1
- nextgen-gallery 4.1.3 → 4.2.0
- really-simple-ssl 9.5.9 → 9.5.11
- simple-embed-code 2.5.2 → 2.6
- woocommerce 10.6.2 → 10.7.0
- woocommerce-gateway-stripe 10.5.3 → 10.6.1
- woocommerce-payments 10.6.0 → 10.7.1
- woocommerce-paypal-payments 3.4.0 → 4.0.3 (major)
- woocommerce-services 3.5.2 → 3.6.1
- wordpress-seo 27.3 → 27.5
- wp-fastest-cache 1.4.7 → 1.4.8
- wp-mail-smtp + pro 4.7.1 → 4.8.0

Excluded:
- google-listings-and-ads (gitignored)
- monpilates-planning, monpilates-giftcards (in-house, edited locally)
- mu-plugins (already in sync with local)

+++ app/public/wp-content/plugins/wp-mail-smtp-pro/src/Pro/Providers/Outlook/OneClick/Auth/Response.php
@@ -60,7 +60,7 @@ class Response {
 	 */
 	public function get_status_code() {
 
-		return wp_remote_retrieve_response_code( $this->response );
+		return (int) wp_remote_retrieve_response_code( $this->response );
 	}
 
 	/**
